requirements.txt
# Core
numpy>=1.24
pandas>=1.5
scikit-learn>=1.3
# joblib>=1.2
pyyaml>=6.0

# Text Preprocessing & Feature Extraction
nltk>=3.8
spacy>=3.7
textblob>=0.17
tqdm>=4.66

# Embeddings
sentence-transformers>=2.2
transformers>=4.40
torch>=2.2

# Visualization (optional but helpful)
matplotlib>=3.7
seaborn>=0.12

# Dev and Utilities
jupyter
python-dotenv

# Additional NLP Libraries
langdetect
stanza
transformers 
sentencepiece 